Separate serialization into a different file
