+++ waspc/e2e-test/test-outputs/waspBuild-golden/expected-files.manifest
@@ -30,7 +30,8 @@ waspBuild/.wasp/build/sdk/wasp/client/router/linkHelpers.ts
 waspBuild/.wasp/build/sdk/wasp/client/router/types.ts
 waspBuild/.wasp/build/sdk/wasp/client/test/index.ts
 waspBuild/.wasp/build/sdk/wasp/client/test/vitest/helpers.tsx
-waspBuild/.wasp/build/sdk/wasp/core/serialization.ts
+waspBuild/.wasp/build/sdk/wasp/core/serialization/base.ts
+waspBuild/.wasp/build/sdk/wasp/core/serialization/index.ts
 waspBuild/.wasp/build/sdk/wasp/core/storage.ts
 waspBuild/.wasp/build/sdk/wasp/dev/index.ts
 waspBuild/.wasp/build/sdk/wasp/dist/api/events.d.ts
@@ -125,10 +126,14 @@ waspBuild/.wasp/build/sdk/wasp/dist/client/test/vitest/helpers.d.ts
 waspBuild/.wasp/build/sdk/wasp/dist/client/test/vitest/helpers.d.ts.map
 waspBuild/.wasp/build/sdk/wasp/dist/client/test/vitest/helpers.jsx
 waspBuild/.wasp/build/sdk/wasp/dist/client/test/vitest/helpers.jsx.map
-waspBuild/.wasp/build/sdk/wasp/dist/core/serialization.d.ts
-waspBuild/.wasp/build/sdk/wasp/dist/core/serialization.d.ts.map
-waspBuild/.wasp/build/sdk/wasp/dist/core/serialization.js
-waspBuild/.wasp/build/sdk/wasp/dist/core/serialization.js.map
+waspBuild/.wasp/build/sdk/wasp/dist/core/serialization/base.d.ts
+waspBuild/.wasp/build/sdk/wasp/dist/core/serialization/base.d.ts.map
+waspBuild/.wasp/build/sdk/wasp/dist/core/serialization/base.js
+waspBuild/.wasp/build/sdk/wasp/dist/core/serialization/base.js.map
+waspBuild/.wasp/build/sdk/wasp/dist/core/serialization/index.d.ts
+waspBuild/.wasp/build/sdk/wasp/dist/core/serialization/index.d.ts.map
+waspBuild/.wasp/build/sdk/wasp/dist/core/serialization/index.js
+waspBuild/.wasp/build/sdk/wasp/dist/core/serialization/index.js.map
 waspBuild/.wasp/build/sdk/wasp/dist/core/storage.d.ts
 waspBuild/.wasp/build/sdk/wasp/dist/core/storage.d.ts.map
 waspBuild/.wasp/build/sdk/wasp/dist/core/storage.js
@@ -344,7 +349,8 @@ waspBuild/.wasp/out/sdk/wasp/client/router/linkHelpers.ts
 waspBuild/.wasp/out/sdk/wasp/client/router/types.ts
 waspBuild/.wasp/out/sdk/wasp/client/test/index.ts
 waspBuild/.wasp/out/sdk/wasp/client/test/vitest/helpers.tsx
-waspBuild/.wasp/out/sdk/wasp/core/serialization.ts
+waspBuild/.wasp/out/sdk/wasp/core/serialization/base.ts
+waspBuild/.wasp/out/sdk/wasp/core/serialization/index.ts
 waspBuild/.wasp/out/sdk/wasp/core/storage.ts
 waspBuild/.wasp/out/sdk/wasp/dev/index.ts
 waspBuild/.wasp/out/sdk/wasp/dist/api/events.d.ts
@@ -439,10 +445,14 @@ waspBuild/.wasp/out/sdk/wasp/dist/client/test/vitest/helpers.d.ts
 waspBuild/.wasp/out/sdk/wasp/dist/client/test/vitest/helpers.d.ts.map
 waspBuild/.wasp/out/sdk/wasp/dist/client/test/vitest/helpers.jsx
 waspBuild/.wasp/out/sdk/wasp/dist/client/test/vitest/helpers.jsx.map
-waspBuild/.wasp/out/sdk/wasp/dist/core/serialization.d.ts
-waspBuild/.wasp/out/sdk/wasp/dist/core/serialization.d.ts.map
-waspBuild/.wasp/out/sdk/wasp/dist/core/serialization.js
-waspBuild/.wasp/out/sdk/wasp/dist/core/serialization.js.map
+waspBuild/.wasp/out/sdk/wasp/dist/core/serialization/base.d.ts
+waspBuild/.wasp/out/sdk/wasp/dist/core/serialization/base.d.ts.map
+waspBuild/.wasp/out/sdk/wasp/dist/core/serialization/base.js
+waspBuild/.wasp/out/sdk/wasp/dist/core/serialization/base.js.map
+waspBuild/.wasp/out/sdk/wasp/dist/core/serialization/index.d.ts
+waspBuild/.wasp/out/sdk/wasp/dist/core/serialization/index.d.ts.map
+waspBuild/.wasp/out/sdk/wasp/dist/core/serialization/index.js
+waspBuild/.wasp/out/sdk/wasp/dist/core/serialization/index.js.map
 waspBuild/.wasp/out/sdk/wasp/dist/core/storage.d.ts
 waspBuild/.wasp/out/sdk/wasp/dist/core/storage.d.ts.map
 waspBuild/.wasp/out/sdk/wasp/dist/core/storage.js
