+++ waspc/e2e-test/test-outputs/waspCompile-golden/expected-files.manifest
@@ -28,7 +28,8 @@ waspCompile/.wasp/out/sdk/wasp/client/router/linkHelpers.ts
 waspCompile/.wasp/out/sdk/wasp/client/router/types.ts
 waspCompile/.wasp/out/sdk/wasp/client/test/index.ts
 waspCompile/.wasp/out/sdk/wasp/client/test/vitest/helpers.tsx
-waspCompile/.wasp/out/sdk/wasp/core/serialization.ts
+waspCompile/.wasp/out/sdk/wasp/core/serialization/base.ts
+waspCompile/.wasp/out/sdk/wasp/core/serialization/index.ts
 waspCompile/.wasp/out/sdk/wasp/core/storage.ts
 waspCompile/.wasp/out/sdk/wasp/dev/index.ts
 waspCompile/.wasp/out/sdk/wasp/dist/api/events.d.ts
@@ -123,10 +124,14 @@ waspCompile/.wasp/out/sdk/wasp/dist/client/test/vitest/helpers.d.ts
 waspCompile/.wasp/out/sdk/wasp/dist/client/test/vitest/helpers.d.ts.map
 waspCompile/.wasp/out/sdk/wasp/dist/client/test/vitest/helpers.jsx
 waspCompile/.wasp/out/sdk/wasp/dist/client/test/vitest/helpers.jsx.map
-waspCompile/.wasp/out/sdk/wasp/dist/core/serialization.d.ts
-waspCompile/.wasp/out/sdk/wasp/dist/core/serialization.d.ts.map
-waspCompile/.wasp/out/sdk/wasp/dist/core/serialization.js
-waspCompile/.wasp/out/sdk/wasp/dist/core/serialization.js.map
+waspCompile/.wasp/out/sdk/wasp/dist/core/serialization/base.d.ts
+waspCompile/.wasp/out/sdk/wasp/dist/core/serialization/base.d.ts.map
+waspCompile/.wasp/out/sdk/wasp/dist/core/serialization/base.js
+waspCompile/.wasp/out/sdk/wasp/dist/core/serialization/base.js.map
+waspCompile/.wasp/out/sdk/wasp/dist/core/serialization/index.d.ts
+waspCompile/.wasp/out/sdk/wasp/dist/core/serialization/index.d.ts.map
+waspCompile/.wasp/out/sdk/wasp/dist/core/serialization/index.js
+waspCompile/.wasp/out/sdk/wasp/dist/core/serialization/index.js.map
 waspCompile/.wasp/out/sdk/wasp/dist/core/storage.d.ts
 waspCompile/.wasp/out/sdk/wasp/dist/core/storage.d.ts.map
 waspCompile/.wasp/out/sdk/wasp/dist/core/storage.js
