+++ waspc/e2e-test/test-outputs/waspComplexTest-golden/expected-files.manifest
@@ -61,7 +61,9 @@ waspComplexTest/.wasp/out/sdk/wasp/client/router/types.ts
 waspComplexTest/.wasp/out/sdk/wasp/client/test/index.ts
 waspComplexTest/.wasp/out/sdk/wasp/client/test/vitest/helpers.tsx
 waspComplexTest/.wasp/out/sdk/wasp/core/auth.ts
-waspComplexTest/.wasp/out/sdk/wasp/core/serialization.ts
+waspComplexTest/.wasp/out/sdk/wasp/core/serialization/base.ts
+waspComplexTest/.wasp/out/sdk/wasp/core/serialization/index.ts
+waspComplexTest/.wasp/out/sdk/wasp/core/serialization/prisma.ts
 waspComplexTest/.wasp/out/sdk/wasp/core/stitches.config.ts
 waspComplexTest/.wasp/out/sdk/wasp/core/storage.ts
 waspComplexTest/.wasp/out/sdk/wasp/dev/index.ts
@@ -281,10 +283,18 @@ waspComplexTest/.wasp/out/sdk/wasp/dist/core/auth.d.ts
 waspComplexTest/.wasp/out/sdk/wasp/dist/core/auth.d.ts.map
 waspComplexTest/.wasp/out/sdk/wasp/dist/core/auth.js
 waspComplexTest/.wasp/out/sdk/wasp/dist/core/auth.js.map
-waspComplexTest/.wasp/out/sdk/wasp/dist/core/serialization.d.ts
-waspComplexTest/.wasp/out/sdk/wasp/dist/core/serialization.d.ts.map
-waspComplexTest/.wasp/out/sdk/wasp/dist/core/serialization.js
-waspComplexTest/.wasp/out/sdk/wasp/dist/core/serialization.js.map
+waspComplexTest/.wasp/out/sdk/wasp/dist/core/serialization/base.d.ts
+waspComplexTest/.wasp/out/sdk/wasp/dist/core/serialization/base.d.ts.map
+waspComplexTest/.wasp/out/sdk/wasp/dist/core/serialization/base.js
+waspComplexTest/.wasp/out/sdk/wasp/dist/core/serialization/base.js.map
+waspComplexTest/.wasp/out/sdk/wasp/dist/core/serialization/index.d.ts
+waspComplexTest/.wasp/out/sdk/wasp/dist/core/serialization/index.d.ts.map
+waspComplexTest/.wasp/out/sdk/wasp/dist/core/serialization/index.js
+waspComplexTest/.wasp/out/sdk/wasp/dist/core/serialization/index.js.map
+waspComplexTest/.wasp/out/sdk/wasp/dist/core/serialization/prisma.d.ts
+waspComplexTest/.wasp/out/sdk/wasp/dist/core/serialization/prisma.d.ts.map
+waspComplexTest/.wasp/out/sdk/wasp/dist/core/serialization/prisma.js
+waspComplexTest/.wasp/out/sdk/wasp/dist/core/serialization/prisma.js.map
 waspComplexTest/.wasp/out/sdk/wasp/dist/core/stitches.config.d.ts
 waspComplexTest/.wasp/out/sdk/wasp/dist/core/stitches.config.d.ts.map
 waspComplexTest/.wasp/out/sdk/wasp/dist/core/stitches.config.js
