+++ waspc/e2e-test/test-outputs/waspJob-golden/expected-files.manifest
@@ -29,7 +29,8 @@ waspJob/.wasp/out/sdk/wasp/client/router/linkHelpers.ts
 waspJob/.wasp/out/sdk/wasp/client/router/types.ts
 waspJob/.wasp/out/sdk/wasp/client/test/index.ts
 waspJob/.wasp/out/sdk/wasp/client/test/vitest/helpers.tsx
-waspJob/.wasp/out/sdk/wasp/core/serialization.ts
+waspJob/.wasp/out/sdk/wasp/core/serialization/base.ts
+waspJob/.wasp/out/sdk/wasp/core/serialization/index.ts
 waspJob/.wasp/out/sdk/wasp/core/storage.ts
 waspJob/.wasp/out/sdk/wasp/dev/index.ts
 waspJob/.wasp/out/sdk/wasp/dist/api/events.d.ts
@@ -124,10 +125,14 @@ waspJob/.wasp/out/sdk/wasp/dist/client/test/vitest/helpers.d.ts
 waspJob/.wasp/out/sdk/wasp/dist/client/test/vitest/helpers.d.ts.map
 waspJob/.wasp/out/sdk/wasp/dist/client/test/vitest/helpers.jsx
 waspJob/.wasp/out/sdk/wasp/dist/client/test/vitest/helpers.jsx.map
-waspJob/.wasp/out/sdk/wasp/dist/core/serialization.d.ts
-waspJob/.wasp/out/sdk/wasp/dist/core/serialization.d.ts.map
-waspJob/.wasp/out/sdk/wasp/dist/core/serialization.js
-waspJob/.wasp/out/sdk/wasp/dist/core/serialization.js.map
+waspJob/.wasp/out/sdk/wasp/dist/core/serialization/base.d.ts
+waspJob/.wasp/out/sdk/wasp/dist/core/serialization/base.d.ts.map
+waspJob/.wasp/out/sdk/wasp/dist/core/serialization/base.js
+waspJob/.wasp/out/sdk/wasp/dist/core/serialization/base.js.map
+waspJob/.wasp/out/sdk/wasp/dist/core/serialization/index.d.ts
+waspJob/.wasp/out/sdk/wasp/dist/core/serialization/index.d.ts.map
+waspJob/.wasp/out/sdk/wasp/dist/core/serialization/index.js
+waspJob/.wasp/out/sdk/wasp/dist/core/serialization/index.js.map
 waspJob/.wasp/out/sdk/wasp/dist/core/storage.d.ts
 waspJob/.wasp/out/sdk/wasp/dist/core/storage.d.ts.map
 waspJob/.wasp/out/sdk/wasp/dist/core/storage.js
