+++ waspc/e2e-test/test-outputs/waspMigrate-golden/expected-files.manifest
@@ -34,7 +34,9 @@ waspMigrate/.wasp/out/sdk/wasp/client/router/linkHelpers.ts
 waspMigrate/.wasp/out/sdk/wasp/client/router/types.ts
 waspMigrate/.wasp/out/sdk/wasp/client/test/index.ts
 waspMigrate/.wasp/out/sdk/wasp/client/test/vitest/helpers.tsx
-waspMigrate/.wasp/out/sdk/wasp/core/serialization.ts
+waspMigrate/.wasp/out/sdk/wasp/core/serialization/base.ts
+waspMigrate/.wasp/out/sdk/wasp/core/serialization/index.ts
+waspMigrate/.wasp/out/sdk/wasp/core/serialization/prisma.ts
 waspMigrate/.wasp/out/sdk/wasp/core/storage.ts
 waspMigrate/.wasp/out/sdk/wasp/dev/index.ts
 waspMigrate/.wasp/out/sdk/wasp/dist/api/events.d.ts
@@ -129,10 +131,18 @@ waspMigrate/.wasp/out/sdk/wasp/dist/client/test/vitest/helpers.d.ts
 waspMigrate/.wasp/out/sdk/wasp/dist/client/test/vitest/helpers.d.ts.map
 waspMigrate/.wasp/out/sdk/wasp/dist/client/test/vitest/helpers.jsx
 waspMigrate/.wasp/out/sdk/wasp/dist/client/test/vitest/helpers.jsx.map
-waspMigrate/.wasp/out/sdk/wasp/dist/core/serialization.d.ts
-waspMigrate/.wasp/out/sdk/wasp/dist/core/serialization.d.ts.map
-waspMigrate/.wasp/out/sdk/wasp/dist/core/serialization.js
-waspMigrate/.wasp/out/sdk/wasp/dist/core/serialization.js.map
+waspMigrate/.wasp/out/sdk/wasp/dist/core/serialization/base.d.ts
+waspMigrate/.wasp/out/sdk/wasp/dist/core/serialization/base.d.ts.map
+waspMigrate/.wasp/out/sdk/wasp/dist/core/serialization/base.js
+waspMigrate/.wasp/out/sdk/wasp/dist/core/serialization/base.js.map
+waspMigrate/.wasp/out/sdk/wasp/dist/core/serialization/index.d.ts
+waspMigrate/.wasp/out/sdk/wasp/dist/core/serialization/index.d.ts.map
+waspMigrate/.wasp/out/sdk/wasp/dist/core/serialization/index.js
+waspMigrate/.wasp/out/sdk/wasp/dist/core/serialization/index.js.map
+waspMigrate/.wasp/out/sdk/wasp/dist/core/serialization/prisma.d.ts
+waspMigrate/.wasp/out/sdk/wasp/dist/core/serialization/prisma.d.ts.map
+waspMigrate/.wasp/out/sdk/wasp/dist/core/serialization/prisma.js
+waspMigrate/.wasp/out/sdk/wasp/dist/core/serialization/prisma.js.map
 waspMigrate/.wasp/out/sdk/wasp/dist/core/storage.d.ts
 waspMigrate/.wasp/out/sdk/wasp/dist/core/storage.d.ts.map
 waspMigrate/.wasp/out/sdk/wasp/dist/core/storage.js
